Refactor: Moved Clean Script to Util
